requirements.txt
py-clob-client>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
eth-account>=0.10.0
web3>=6.0.0
websockets>=12.0
pytest>=7.4.0
